Exponentiation

+++ Code/headerFile.h
@@ -9,6 +9,6 @@ void BubbleAndCountingTask(); // 2
 
 void ExponentiationTask(); // 3
 
-long ExponentiationLogTime(int baseOfDegree, int exponent);
+long ExponentiationLogTime(long baseOfDegree, long exponent);
 
 void Swap(int *first, int *second);
